requirements.txt
fastapi==0.63.0
cryptography==3.3.1
PyJWT==2.0.0
fastapi-utils==0.2.1
alembic==1.4.3
psycopg2-binary==2.8.6
pydantic==1.7.3
gunicorn==20.0.4
uvicorn[standard]
